bitbake: siggen/runqueue: Allow skipping of rehash if not present on hashserve
If a unihash isn't present on the hashserver, it is extremely unlikely any
following unihashes would be either. Use this information to skip unihash
checks in that scenario where there is low matching to allow the build
to proceed more quickly. This significantly improves performance on the
low match scenario.

(Bitbake rev: b3ebad873ce161ecffb7469ccd3737e17263fcf2)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>